backend/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==4.3.0
flask-pytest==0.0.5
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.5.2
requests==2.21.0
PyMySQL==0.9.3
gunicorn==20.0.4
